third_party/blink/web_tests/platform/generic/fast/forms/select/change-multiple-preserve-selection-expected.txt

1
2
3
 
1
2
3
 
1
2
3
 
1
2
3
 
1
2
3
PASS successfullyParsed is true

TEST COMPLETE

PASS selectElement.selectedIndex is -1
PASS selectElement.selectedIndex is 0
PASS selectElement.selectedIndex is 0
PASS selectElement.selectedIndex is -1
PASS selectElement.selectedIndex is 0
PASS selectElement.selectedIndex is 0
PASS selectElement.options[0].selected is true
PASS selectElement.options[1].selected is true
PASS selectElement.options[2].selected is false
PASS selectElement.options[0].selected is true
PASS selectElement.options[1].selected is true
PASS selectElement.options[2].selected is false
PASS selectElement.options[0].selected is true
PASS selectElement.options[1].selected is false
PASS selectElement.options[2].selected is false
